+++ apps/dsp-app/src/app/app.module.ts
@@ -23,6 +23,11 @@ import {
 } from '@dasch-swiss/vre/shared/app-config';
 import { AppDatePickerComponent } from '@dasch-swiss/vre/shared/app-date-picker';
 import { AppErrorHandler } from '@dasch-swiss/vre/shared/app-error-handler';
+import {
+  CreateListInfoPageComponent,
+  ReusableListInfoFormComponent,
+  ListComponents,
+} from '@dasch-swiss/vre/shared/app-list';
 import { NotificationService } from '@dasch-swiss/vre/shared/app-notification';
 import {
   AppProgressIndicatorComponent,
@@ -40,6 +45,7 @@ import {
   MultiLanguageTextareaComponent,
   MutiLanguageInputComponent,
 } from '@dasch-swiss/vre/shared/app-string-literal';
+import { UiComponents } from '@dasch-swiss/vre/shared/app-ui';
 import { TimeFormatDirective } from '@dsp-app/src/app/project/time-input/time-format.directive';
 import { TimeInputComponent } from '@dsp-app/src/app/project/time-input/time-input.component';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
@@ -52,7 +58,6 @@ import { ColorPickerModule } from 'ngx-color-picker';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ConfirmDialogComponent } from './main/action/confirm-dialog/confirm-dialog.component';
 import { HintComponent } from './main/action/hint/hint.component';
 import { LoginFormComponent } from './main/action/login-form/login-form.component';
 import { SelectedResourcesComponent } from './main/action/selected-resources/selected-resources.component';
@@ -90,17 +95,6 @@ import { EditProjectFormPageComponent } from './project/edit-project-form-page/e
 import { ImageDisplayAbsoluteComponent } from './project/image-settings/image-display-absolute.component';
 import { ImageDisplayRatioComponent } from './project/image-settings/image-display-ratio.component';
 import { ImageSettingsComponent } from './project/image-settings/image-settings.component';
-import { ActionBubbleComponent } from './project/list/action-bubble/action-bubble.component';
-import { ListItemComponent } from './project/list/list-item/list-item.component';
-import { ListItemElementComponent } from './project/list/list-item-element/list-item-element.component';
-import { CreateListItemDialogComponent } from './project/list/list-item-form/edit-list-item/create-list-item-dialog.component';
-import { EditListItemDialogComponent } from './project/list/list-item-form/edit-list-item/edit-list-item-dialog.component';
-import { ListItemFormComponent } from './project/list/list-item-form/list-item-form.component';
-import { ReusableListItemFormComponent } from './project/list/list-item-form/reusable-list-item-form.component';
-import { ListComponent } from './project/list/list.component';
-import { CreateListInfoPageComponent } from './project/list/reusable-list-info-form/create-list-info-page.component';
-import { EditListInfoDialogComponent } from './project/list/reusable-list-info-form/edit-list-info-dialog.component';
-import { ReusableListInfoFormComponent } from './project/list/reusable-list-info-form/reusable-list-info-form.component';
 import { CreateResourceClassDialogComponent } from './project/ontology/create-resource-class-dialog/create-resource-class-dialog.component';
 import { EditResourceClassDialogComponent } from './project/ontology/edit-resource-class-dialog/edit-resource-class-dialog.component';
 import { OntologyFormComponent } from './project/ontology/ontology-form/ontology-form.component';
@@ -167,8 +161,9 @@ export function httpLoaderFactory(httpClient: HttpClient) {
     ...ResourcePageComponents,
     ...RepresentationsComponents,
     ...CommonToMoveComponents,
+    ...ListComponents,
+    ...UiComponents,
     AccountComponent,
-    ActionBubbleComponent,
     AddUserComponent,
     AdminImageDirective,
     AdvancedSearchContainerComponent,
@@ -180,7 +175,6 @@ export function httpLoaderFactory(httpClient: HttpClient) {
     ColorPickerComponent,
     CommonInputComponent,
     ComparisonComponent,
-    ConfirmDialogComponent,
     CookiePolicyComponent,
     CreateUserDialogComponent,
     CreateResourceClassDialogComponent,
@@ -205,15 +199,8 @@ export function httpLoaderFactory(httpClient: HttpClient) {
     KnoraDatePipe,
     LoadingButtonDirective,
     LinkifyPipe,
-    ListComponent,
-    EditListInfoDialogComponent,
-    EditListItemDialogComponent,
     CreateListInfoPageComponent,
     ReusableListInfoFormComponent,
-    ReusableListItemFormComponent,
-    ListItemComponent,
-    ListItemElementComponent,
-    ListItemFormComponent,
     ListViewComponent,
     LoginFormComponent,
     MembershipComponent,
@@ -267,7 +254,6 @@ export function httpLoaderFactory(httpClient: HttpClient) {
     ProjectTileComponent,
     DataModelsComponent,
     IsFalsyPipe,
-    CreateListItemDialogComponent,
     ImageDisplayAbsoluteComponent,
     MultipleViewerComponent,
   ],