+++ libs/vre/shared/app-common/src/lib/resource.util.ts
@@ -1,0 +1,13 @@
+import { PermissionUtil, ReadResource, ReadValue } from '@dasch-swiss/dsp-js';
+
+export class ResourceUtil {
+  public static userCanEdit(instance: ReadResource | ReadValue) {
+    const permissions = PermissionUtil.allUserPermissions(instance.userHasPermission as 'RV' | 'V' | 'M' | 'D' | 'CR');
+    return permissions.includes(PermissionUtil.Permissions.M);
+  }
+
+  public static userCanDelete(instance: ReadResource | ReadValue) {
+    const permissions = PermissionUtil.allUserPermissions(instance.userHasPermission as 'RV' | 'V' | 'M' | 'D' | 'CR');
+    return permissions.includes(PermissionUtil.Permissions.D);
+  }
+}