+++ libs/vre/shared/app-config/src/lib/app-config/dsp-dialog-config.ts
@@ -3,7 +3,7 @@ import { MatDialogConfig } from '@angular/material/dialog';
 export class DspDialogConfig {
   static dialogDrawerConfig<T = any>(dialogData?: T): MatDialogConfig<T> {
     return {
-      minHeight: 'calc(100vh)',
+      height: 'calc(100vh)',
       minWidth: 600,
       maxWidth: '100%',
       position: {