+++ libs/vre/shared/app-list/src/list.components.ts
@@ -1,0 +1,25 @@
+import { ActionBubbleComponent } from './lib/action-bubble/action-bubble.component';
+import { ListItemComponent } from './lib/list-item/list-item.component';
+import { ListItemElementComponent } from './lib/list-item-element/list-item-element.component';
+import { CreateListItemDialogComponent } from './lib/list-item-form/edit-list-item/create-list-item-dialog.component';
+import { EditListItemDialogComponent } from './lib/list-item-form/edit-list-item/edit-list-item-dialog.component';
+import { ListItemFormComponent } from './lib/list-item-form/list-item-form.component';
+import { ReusableListItemFormComponent } from './lib/list-item-form/reusable-list-item-form.component';
+import { ListComponent } from './lib/list.component';
+import { CreateListInfoPageComponent } from './lib/reusable-list-info-form/create-list-info-page.component';
+import { EditListInfoDialogComponent } from './lib/reusable-list-info-form/edit-list-info-dialog.component';
+import { ReusableListInfoFormComponent } from './lib/reusable-list-info-form/reusable-list-info-form.component';
+
+export const ListComponents = [
+  ListItemComponent,
+  ListItemElementComponent,
+  ListItemFormComponent,
+  ReusableListItemFormComponent,
+  CreateListItemDialogComponent,
+  EditListItemDialogComponent,
+  CreateListInfoPageComponent,
+  EditListInfoDialogComponent,
+  ReusableListInfoFormComponent,
+  ActionBubbleComponent,
+  ListComponent,
+];