+++ libs/vre/shared/app-resource-properties/src/lib/switch-components/link-switch.component.ts
@@ -36,7 +36,7 @@ export class LinkSwitchComponent implements IsSwitchComponent {
   }
 
   get link() {
-    return `/resource${this._resourceService.getResourcePath(this.control.value)}`;
+    return this.control.value ? `/resource${this._resourceService.getResourcePath(this.control.value)}` : '#';
   }
 
   constructor(private _resourceService: ResourceService) {}